# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 学生综合素质测评管理系统 · 服务端 API
  version: 1.0.0
  summary: 服务端对外接口的唯一权威契约
  description: |
    本文件是服务端接口的**唯一权威契约**（决策 §1.3 契约形式）。
    `server/web`（Fastify + PostgreSQL）与 `server/cloudflare`（Workers + D1）
    两套实现必须逐字段对齐本契约，一致性由 `contract-tests/` 的同一断言集保证。

    ## 通用约定

    - 路径版本化：`/api/v1/*`。
    - 响应统一包裹：成功 `{ "ok": true, "data": ... }`，失败 `{ "ok": false, "error": "中文描述" }`。
    - 错误文案一律为**简体中文**（管理端 `server-client.ts` 直接透传给用户）。
    - 时间戳统一为 epoch 毫秒整数。
    - 服务端**不存储**分数明细、证明材料与任何私钥（数据主权原则）。

    ## 鉴权

    | 角色 | 方式 | 说明 |
    |------|------|------|
    | 管理端写接口 | `Authorization: Bearer <unitToken>` | 激活时下发，绑定 unitId + installId，随授权码作废/换机失效（决策 #29） |
    | 管理端标识头 | `X-Unit-Id` / `X-Install-Id` | 辅助标识与审计，**不作为信任凭证** |
    | 学生端 | 路径中的 `applyId` | 122 bit 随机 UUID，服务端只存 SHA-256 哈希 + 严格限流（决策 #40） |
    | 管理后台 | `Authorization: Bearer <accessToken>` | 15 分钟短期 token，配可删除的刷新 token（决策 #40） |

    ## 状态机

    - 批次：`draft → active → closed`，单向推进。
    - 申请：单轮内 `draft → submitted → imported → reviewing → confirmed`，**不允许回退**；
      整班重审由一级发起新一轮（`review_round + 1`，从 `reviewing` 开始，决策 #30/#31）。
  contact:
    name: 项目文档
    url: https://github.com/local/dys-management-system/blob/main/docs/server-plan.md
servers:
  - url: http://127.0.0.1:3100
    description: 本地开发（server/web）
  - url: http://127.0.0.1:8787
    description: 本地开发（server/cloudflare，wrangler dev）
  - url: https://{host}
    description: 生产部署
    variables:
      host:
...
security:
  - unitToken: []
components:
  securitySchemes:
    unitToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 单位令牌，激活时由 `POST /api/v1/authorize` 下发（决策 #29）。
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 管理后台访问令牌，有效期 15 分钟（决策 #40）。
  parameters:
...
    unitIdPath:
      name: unitId
      in: path
      required: true
      description: 单位 id。
      schema: { $ref: '#/components/schemas/UnitId' }
...
  schemas:
    EpochMs:
      description: epoch 毫秒时间戳。
      type: integer
      minimum: 0
      examples: [1755500000000]
    UnitId:
      description: 单位 id（camelCase，与学生端单位树 value 一致）。
      type: string
      pattern: '^[a-zA-Z][a-zA-Z0-9]*$'
      maxLength: 64
      examples: ['nxuLx']
...
    LicenseCode:
      description: 授权码（大写字母数字分组）。
      type: string
      pattern: '^[A-Z0-9]{4}(-[A-Z0-9]{4}){3}$'
      examples: ['A1B2-C3D4-E5F6-G7H8']
...
    UnitSummary:
      description: 单位摘要（激活响应与后台列表共用）。
      type: object
      required: [unitId, unitName, unitType]
      properties:
        unitId: { $ref: '#/components/schemas/UnitId' }
        unitName: { type: string, examples: ['励行书院'] }
        unitType:
          enum: [college, department, other]
          examples: ['college']
        level:
          description: 单位层级：1 一级（学校）/ 2 二级（书院）。
          enum: [1, 2]
        parentUnitId:
          description: 一级单位 id（二级单位必有）。
          oneOf:
            - { $ref: '#/components/schemas/UnitId' }
            - { type: 'null' }
      additionalProperties: false
...
    ErrorBody:
      description: 统一失败响应。
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          description: 面向用户的中文错误描述（管理端直接透传）。
          type: string
          minLength: 1
      additionalProperties: false
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
          examples:
            default:
              value: { ok: false, error: '请求参数不合法' }
    Unauthorized:
      description: 未认证（缺少或无效令牌）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
          examples:
            default:
              value: { ok: false, error: '未登录或登录已过期' }
...
    ServerError:
      description: 服务端异常
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
          examples:
            default:
              value: { ok: false, error: '服务端异常，请稍后重试' }
paths:
...
  /api/v1/admin/units:
    get:
      tags: [后台单位与授权]
      summary: 单位列表（接口 18）
      description: 返回全部单位（含一二级层级关系）。
      operationId: adminListUnits
      security:
        - adminToken: []
      parameters:
        - name: level
          in: query
          required: false
          description: 按层级过滤。
          schema: { enum: [1, 2] }
      responses:
        '200':
          description: 单位列表
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [units]
                    properties:
                      units:
                        type: array
                        items: { $ref: '#/components/schemas/UnitSummary' }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [后台单位与授权]
      summary: 创建单位（接口 18）
      description: |
        创建单位。二级单位必须指定 `parentId` 与 `templateId`（绑定配置模板），
        **创建即签发首个授权码**并返回。落审计日志。
      operationId: adminCreateUnit
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [unitId, name, level]
              properties:
                unitId: { $ref: '#/components/schemas/UnitId' }
                name: { type: string, minLength: 1, maxLength: 64 }
                level: { enum: [1, 2] }
                unitType:
                  enum: [college, department, other]
                parentId:
                  description: 二级单位必填。
                  $ref: '#/components/schemas/UnitId'
                templateId:
                  description: 二级单位必填，绑定的配置模板 id。
                  type: string
                licenseMonths:
                  description: 首个授权码有效月数，缺省 12。
                  type: integer
                  minimum: 1
                  maximum: 120
              additionalProperties: false
      responses:
        '200':
          description: 创建成功（含首个授权码）
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [unit]
                    properties:
                      unit: { $ref: '#/components/schemas/UnitSummary' }
                      code:
                        description: 一级单位无授权码时为 null。
                        oneOf:
                          - { $ref: '#/components/schemas/LicenseCode' }
                          - { type: 'null' }
                      expiresAt:
                        oneOf: [{ $ref: '#/components/schemas/EpochMs' }, { type: 'null' }]
                    additionalProperties: false
                additionalProperties: false
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: unitId 已存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                duplicate:
                  value: { ok: false, error: '该单位标识已存在' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/units/{unitId}:
    get:
      tags: [后台单位与授权]
      summary: 单位详情（接口 18）
      description: 返回单位、绑定的配置模板版本与授权码列表。
      operationId: adminGetUnit
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/unitIdPath'
      responses:
        '200':
          description: 单位详情
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [unit, licenses]
                    properties:
                      unit: { $ref: '#/components/schemas/UnitSummary' }
                      publicKeyJwk:
                        oneOf: [{ $ref: '#/components/schemas/Jwk' }, { type: 'null' }]
                      template:
                        oneOf:
                          - { $ref: '#/components/schemas/TemplateVersion' }
                          - { type: 'null' }
                      licenses:
                        type: array
                        items:
                          type: object
                          required: [code, expiresAt, status, createdAt]
                          properties:
                            code: { $ref: '#/components/schemas/LicenseCode' }
                            expiresAt: { $ref: '#/components/schemas/EpochMs' }
                            status: { $ref: '#/components/schemas/LicenseStatus' }
                            renewCount: { type: integer, minimum: 0 }
                            createdAt: { $ref: '#/components/schemas/EpochMs' }
                          additionalProperties: false
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      tags: [后台单位与授权]
      summary: 删除单位（接口 18）
      description: |
        删除指定单位及其全部关联数据（授权码、单位令牌、换机记录、批次、申请状态、
        配置模板，均 `ON DELETE CASCADE`）。一级单位存在下级单位时拒绝删除（409）。
        落审计日志。
      operationId: adminDeleteUnit
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/unitIdPath'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [ok]
                    properties:
                      ok: { const: true }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 一级单位下仍有下级单位
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                hasChildren:
                  value: { ok: false, error: '该单位下仍有下级单位，请先删除下级单位' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/units/{unitId}/licenses:
    post:
      tags: [后台单位与授权]
      summary: 签发授权码（接口 19）
      description: 为指定二级单位签发新授权码。落审计日志。
      operationId: adminIssueLicense
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/unitIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [months]
              properties:
                months: { type: integer, minimum: 1, maximum: 120 }
              additionalProperties: false
      responses:
        '200':
          description: 签发成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [code, expiresAt]
                    properties:
                      code: { $ref: '#/components/schemas/LicenseCode' }
                      expiresAt: { $ref: '#/components/schemas/EpochMs' }
                    additionalProperties: false
                additionalProperties: false
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/licenses/{code}/revoke:
    post:
      tags: [后台单位与授权]
      summary: 作废授权码（接口 19）
      description: |
        作废授权码。同时使该授权码派生的全部 `unitToken` 失效（决策 #29）。
        已作废的授权码重复作废幂等。落审计日志。
      operationId: adminRevokeLicense
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/licenseCodePath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 200 }
              additionalProperties: false
      responses:
        '200':
          description: 已作废
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [code, status]
                    properties:
                      code: { $ref: '#/components/schemas/LicenseCode' }
                      status: { const: revoked }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/licenses/{code}/renew:
    post:
      tags: [后台单位与授权]
      summary: 续期授权码（接口 19）
      description: |
        续期授权码（决策 #40）：过期不再强制重新激活，管理端经
        `GET /api/v1/license/status` 即可刷新本地有效期。已作废的授权码不可续期（409）。
        落审计日志。
      operationId: adminRenewLicense
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/licenseCodePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [months]
              properties:
                months: { type: integer, minimum: 1, maximum: 120 }
              additionalProperties: false
            examples:
              default:
                value: { months: 12 }
      responses:
        '200':
          description: 续期成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [code, expiresAt, status, renewCount]
                    properties:
                      code: { $ref: '#/components/schemas/LicenseCode' }
                      expiresAt: { $ref: '#/components/schemas/EpochMs' }
                      status: { $ref: '#/components/schemas/LicenseStatus' }
                      renewCount: { type: integer, minimum: 1 }
                    additionalProperties: false
                additionalProperties: false
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 已作废的授权码不可续期
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                revoked:
                  value: { ok: false, error: '授权码已作废，不可续期' }
        '500': { $ref: '#/components/responses/ServerError' }

  # -------------------------------------------------------- 7. 后台配置模板
  /api/v1/admin/templates:
    get:
      tags: [后台配置模板]
      summary: 配置模板列表（接口 20）
      description: 返回每个模板 id 的最新版本行（含 draft）。
      operationId: adminListTemplates
      security:
        - adminToken: []
      responses:
        '200':
          description: 模板列表
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [templates]
                    properties:
                      templates:
                        type: array
                        items: { $ref: '#/components/schemas/TemplateVersion' }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [后台配置模板]
      summary: 上传配置模板（接口 20）
      description: |
        上传单位配置 JSON（须通过 `unit-config.schema.json` 校验）。
        以 `draft` 状态写入 (id, version, revision)；同三元组已存在 `published` 版本时返回 409
        （已发布不可修改，决策 #37）。落审计日志。
      operationId: adminUploadTemplate
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: { $ref: '#/components/schemas/UnitConfig' }
              additionalProperties: false
      responses:
        '200':
          description: 上传成功（draft）
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data: { $ref: '#/components/schemas/TemplateVersion' }
                additionalProperties: false
        '400':
          description: 配置未通过 schema 校验
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                invalid:
                  value: { ok: false, error: '配置格式不合法：dyf.categories 不能为空' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: 该版本已发布，不可覆盖
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                published:
                  value: { ok: false, error: '该版本已发布，请升修订号后再上传' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/templates/{templateId}/versions:
    get:
      tags: [后台配置模板]
      summary: 模板版本历史（接口 20）
      description: 按 (version, revision) 倒序返回该模板的全部历史版本（决策 #37）。
      operationId: adminListTemplateVersions
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/templateIdPath'
      responses:
        '200':
          description: 版本历史
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [versions]
                    properties:
                      versions:
                        type: array
                        items: { $ref: '#/components/schemas/TemplateVersion' }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/templates/{templateId}/versions/{version}/{revision}/config:
    get:
      tags: [后台配置模板]
      summary: 查看模板配置内容（接口 20）
      description: |
        返回指定版本的完整单位配置 JSON（后台只读查看）。
        结构与上传时的 `UnitConfig` 完全一致。
      operationId: adminGetTemplateConfig
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/templateIdPath'
        - name: version
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
        - name: revision
          in: path
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: 配置内容
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [config]
                    properties:
                      config: { $ref: '#/components/schemas/UnitConfig' }
                    additionalProperties: false
                additionalProperties: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/admin/templates/{templateId}/publish:
    post:
      tags: [后台配置模板]
      summary: 发布模板版本（接口 20）
      description: |
        将指定 (version, revision) 置为 `published`，此后不可编辑；
        同 id 的其他 `published` 版本转为 `archived`。已创建的批次按快照三元组引用，不受影响。
        落审计日志。
      operationId: adminPublishTemplate
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/templateIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, revision]
              properties:
                version: { type: integer, minimum: 1 }
                revision: { type: integer, minimum: 0 }
              additionalProperties: false
      responses:
        '200':
          description: 已发布
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data: { $ref: '#/components/schemas/TemplateVersion' }
                additionalProperties: false
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 该版本已是 published 或已 archived
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
              examples:
                archived:
                  value: { ok: false, error: '已归档的版本不可发布' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ------------------------------------------------------------ 8. 后台运维
...